Throw AssertionError if a test overwrites its own test files.
If you write the same test file in a test you won't compile the first one and so won't actually test it.

This CL fixes adds the exception, fixes tests which violate it and fixes the FormatString checker which was actually not WAI. It was a minor issue whereby calls (such as readLine) on the Console object were not being checked.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=259902914

+++ test_helpers/src/main/java/com/google/errorprone/ErrorProneInMemoryFileManager.java
@@ -121,6 +121,10 @@ public class ErrorProneInMemoryFileManager extends JavacFileManager {
     Path path = resolvePath(fileName);
     try {
       Files.createDirectories(path.getParent());
+      if (Files.exists(path)) {
+        throw new AssertionError(
+            "Refusing to overwrite an existing file (all results from that file would be lost)");
+      }
       Files.write(path, Joiner.on('\n').join(lines).getBytes(UTF_8));
     } catch (IOException e) {
       throw new AssertionError(e);
